+++ common/settings/user_settings.py
@@ -7,8 +7,17 @@
 #DEBUG=False
 #CacheAge = 86400*30                # max age of cached entries
 #CacheAgeNegative = 86400*5         # max age of negative cache entries
+#CacheType = ""                     # [ file, db ]
 #CacheFile = ""                     # full path to (writable) cache file
-#CacheType = ""                     #
+#CacheDBType = ""                   # cache DB type for CacheType=db
+                                    # [ mysql, mongodb ]
+#CacheDBHost = ""                   # cache DB host for CacheType=db
+#CacheDBPort = ""                   # cache DB port for CacheType=db
+#CacheDBName = ""                   # cache DB name for CacheType=db
+#CacheDBTable = ""                  # cache DB table for CacheType=db
+#CacheDBUser = ""                   # cache DB user for CacheType=db
+#CacheDBPassword = ""               # cache DB password for CacheType=db
+#CacheDBAuthDB = ""                 # cache DB auth DB for CacheType=db
 #LogFile = ""                       # full path to (writable) log file
 #LogLevel = "W"                     # don't write logs below this level
 #ConfigFile = ""                    # full path to (readable) config file